remove index_t definition
conflicts with prev code base.

Also std::optional doubles the size of the underlying type.
horrifying!  I don't want to have to give 16 bytes over for something
that could be embedded in the 8 bytes of the u64 directly.  I'm
thinking we could just use i64's instead, sacrificing that top bit to
indicate if the value is present or not.